+++ CHANGELOG.md
@@ -1,10 +1,20 @@
-#  (2021-03-22)
+#  (2021-04-09)
 
 
 ### Features
 
+* **analysis:** add analysis & api ([84eb406](https://github.com/ibwei/umi-base/commit/84eb40619b71b9f94881d9f7a927bc9ea0133036))
+* **analysis:** 初步完成所有的逻辑 ([689231d](https://github.com/ibwei/umi-base/commit/689231d4b0762381c1547307ea5b4c2f0a6a1483))
+* **answer:** handle answer ([d484e79](https://github.com/ibwei/umi-base/commit/d484e797271d8f56ead01b8aab1683bacb39bfda))
 * **dva:** 初步完善好 dva 部分 ([6af32a3](https://github.com/ibwei/umi-base/commit/6af32a3ea4aa181cbd80f7c0432157d8d2a9b4a5))
+* **home:** 初步完成首页 ([02bf3bc](https://github.com/ibwei/umi-base/commit/02bf3bc788e0fb04960d8c03384dd5e9674719b4))
+* **husky:** add husky ([211ceb6](https://github.com/ibwei/umi-base/commit/211ceb6c7f24aac0face2584b3605af630cf79cf))
+* **husky:** add husky ([bdee796](https://github.com/ibwei/umi-base/commit/bdee796846e0847125dd1bd2dfd069f26a9a5457))
+* **login:** add login & plugins ([a980c43](https://github.com/ibwei/umi-base/commit/a980c43986e32ace81aca035487463ad5c35d087))
+* **question:** 初步完成逻辑 ([c3c26b8](https://github.com/ibwei/umi-base/commit/c3c26b81e57c1253d3ab14d460be5f860e7da94c))
+* **question:** 答题页面 ([2ac09ce](https://github.com/ibwei/umi-base/commit/2ac09ce86f8ae7636a64c52813abdd6c7a622d79))
 * **router & git-cz:** 完善路由配置以及添加了git message 工具 ([f36593c](https://github.com/ibwei/umi-base/commit/f36593cd91ac56c243843adca9be79047fb2ed3f))
+* **usehook:** test for useMemo,useCallback ([4240954](https://github.com/ibwei/umi-base/commit/42409549baa49e56fb7ae35b995a90e396070ac5))